# FilesService/swagger.yaml
basePath: /files/
definitions:
    Image:
        description: Image defines a properties of a image file that are used to construct filename
        properties:
            extension:
                maxLength: 8
                minLength: 2
                type: string
                x-go-name: FileExtension
            resolution:
                maxLength: 16
                minLength: 7
                type: string
                x-go-name: Resolution
            type:
                description: Image type determining general image purpose (eg. Albedo, Roughness)
                maxLength: 64
                minLength: 2
                type: string
                x-go-name: ImgType
        required:
            - type
            - resolution
            - extension
        type: object
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/data
    categoryContents:
        properties:
            directories:
                description: List of directories stored inside given category
                items:
                    type: string
                type: array
                x-go-name: Directories
        type: object
        x-go-name: CategoryResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
...
    imageSet:
        description: ImageSet defines a properties of a set of images contributing to an entire texture with various resolutions or image types
        properties:
            category:
                description: Category determining storage subdirectory
                maxLength: 128
                minLength: 3
                type: string
                x-go-name: Category
            id:
                description: ID as it is stored in the database
                maxLength: 64
                minLength: 2
                type: string
                x-go-name: ID
            images:
                description: List of images contriburing to this particular imageSet
                items:
                    $ref: '#/definitions/Image'
                type: array
                x-go-name: Images
        type: object
        x-go-name: ImageSet
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/data
    message:
        properties:
            message:
                type: string
                x-go-name: Message
        type: object
        x-go-name: MessageResponse
        x-go-package: github.com/KrzysztofSieczkiewicz/go--model-viewer-backend/FilesService/response
host: localhost:9090
info:
    description: Documentation of FilesApi
    title: FilesService.
    version: 1.0.0
paths:
...
    /imageCategories/{category}:
        delete:
            consumes:
                - application/json
            description: Delete existing category, requires being emptied beforehand
            operationId: deleteCategory
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        get:
            description: List subdirectories available in the category
            operationId: getCategory
            produces:
                - application/json
            responses:
                "200":
                    description: categoryContents
                    schema:
                        $ref: '#/definitions/categoryContents'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        post:
            description: Creates a requested directory path
            operationId: postCategory
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        put:
            description: Update existing Category
            operationId: putCategory
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
    /imageSets/{category}/{id}:
        delete:
            consumes:
                - application/json
            description: Delete existing imageset
            operationId: deleteImageSet
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        get:
            description: Return ImageSet details and available Images
            operationId: getImageSet
            produces:
                - application/json
            responses:
                "200":
                    description: imageSet
                    schema:
                        $ref: '#/definitions/imageSet'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        post:
            description: Create a new image set
            operationId: postImageSet
            produces:
                - application/json
            responses:
                "201":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
        put:
            consumes:
                - application/json
            description: Update existing imageset id or category. Allows to move imageset to the different category, but it must be initialized beforehand
            operationId: putImageSet
            produces:
                - application/json
            responses:
                "200":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "400":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "403":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "404":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
                "500":
                    description: message
                    schema:
                        $ref: '#/definitions/message'
            tags:
                - imageSets
...
schemes:
    - http
    - https
swagger: "2.0"